requirements.txt
torch
torchvision
Pillow
numpy
streamlit
tqdm
opencv-python
